v1.03

+++ src/test/java/com/commandoby/stringCalculator/GetAnswerTests.java
@@ -1,5 +1,7 @@
 package com.commandoby.stringCalculator;
 
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -12,13 +14,14 @@ public class GetAnswerTests {
 
 	@ParameterizedTest
 	@CsvFileSource(resources = "/csv/GetAnswerTests.csv", numLinesToSkip = 1)
-	public void getAnswer_Test(String input, double expected) throws InvalidCharacterException, SubEquationException {
-		double actualAnswer = Application.getAnswer(input);
+	public void getAnswer_Test(String input, BigDecimal expected)
+			throws InvalidCharacterException, SubEquationException {
+		BigDecimal actualAnswer = Application.getAnswer(input);
 		Assertions.assertEquals(expected, actualAnswer);
 	}
 
-    @AfterAll
-    public static void tearDown() {
-        System.out.println("All GetAnswer tests are finished!");
-    }
+	@AfterAll
+	public static void tearDown() {
+		System.out.println("All GetAnswer tests are finished!");
+	}
 }
